test/e2e/requirements.txt
kafka-python == 2.0.1
pytest == 5.3.5
